refactor: port tests separation to master (#70)
* Refactor: port tests separation to master

* CI: ignore pylint erros
